+++ internal/pfcpsim/export/export.go
@@ -1,4 +1,4 @@
-package fuzz
+package export
 
 import (
 	"errors"
